Edit DemoDrawer to use getPaletteColor
